feat: revamp and adapt the new topbar navigation on pipeline builder
