+++ apps/console/src/pages/[entity]/pipelines/[id]/builder.tsx
@@ -4,8 +4,10 @@ import { PipelineBuilderMainView } from "@instill-ai/toolkit";
 import { ConsoleCorePageHead } from "../../../../components";
 import { NextPageWithLayout } from "../../../_app";
 import { useTrackToken } from "../../../../lib/useTrackToken";
+import { useAccessToken } from "lib/useAccessToken";
 
 const PipelineBuilderPage: NextPageWithLayout = () => {
+  useAccessToken();
   useTrackToken({ enabled: true });
 
   return (
