+++ packages/toolkit/src/lib/store/index.ts
@@ -1,5 +1,4 @@
 export * from "./useConfigureModelFormStore";
-export * from "./useConfigurePipelineFormStore";
 export * from "./useConfigureSourceFormStore";
 export * from "./useCreateResourceFormStore";
 export * from "./useMessageBoxStore";
