+++ packages/toolkit/src/view/pipeline-builder/components/dialogs/index.ts
@@ -1,3 +1,3 @@
 export * from "./publish-pipeline-dialog";
-export * from "./select-component-definition-dialog";
+export * from "./select-component-dialog";
 export * from "./share-pipeline-dialog";
