--- packages/toolkit/src/view/pipeline-builder/components/dialogs/select-component-definition-dialog/index.ts
@@ -1,1 +1,0 @@
-export * from "./SelectPipelineComponentDefinitionDialog";
