+++ packages/toolkit/src/view/pipeline-builder/components/dialogs/select-component-dialog/DialogSection.tsx
