+++ packages/toolkit/src/view/pipeline-builder/components/dialogs/select-component-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from "./SelectComponentDialog";
