+++ packages/toolkit/src/view/pipeline-builder/components/index.ts
@@ -4,7 +4,7 @@ export * from "./CreateResourceDialog";
 export * from "./CustomEdge";
 export * from "./CustomHandle";
 export * from "./PipelineNameForm";
-export * from "./PipelineToolkitModal";
+export * from "./PipelineToolkitDialog";
 export * from "./ReadOnlyPipelineBuilder";
 export * from "./ComponentOutputReferenceHints";
