+++ packages/toolkit/src/view/pipeline-builder/index.ts
@@ -1,6 +1,6 @@
 export * from "./components";
 export * from "./lib";
 export * from "./Flow";
-export * from "./FlowControl";
+export * from "./PipelineBuilderMainView";
 export * from "./RightPanel";
 export * from "./type";
