+++ packages/toolkit/src/view/pipeline/index.ts
@@ -1,5 +1,4 @@
 export * from "./view-pipeline";
 export * from "./view-pipelines";
-export * from "./PipelineBuilderMainView";
 export * from "./PipelineTablePlaceholder";
 export * from "./PipelinesTable";
